skipping deleteOldClusters if gcloud command is not present

+++ Jenkinsfile
@@ -48,23 +48,25 @@ void shutdownCluster(String CLUSTER_SUFFIX) {
 void deleteOldClusters(String FILTER) {
     withCredentials([string(credentialsId: 'GCP_PROJECT_ID', variable: 'GCP_PROJECT'), file(credentialsId: 'gcloud-key-file', variable: 'CLIENT_SECRET_FILE')]) {
         sh """
-            gcloud auth activate-service-account --key-file $CLIENT_SECRET_FILE
-            gcloud config set project $GCP_PROJECT
-            for GKE_CLUSTER in \$(gcloud container clusters list --format='csv[no-heading](name)' --filter="$FILTER"); do
-                GKE_CLUSTER_STATUS=\$(gcloud container clusters list --format='csv[no-heading](status)' --filter="\$GKE_CLUSTER")
-                retry=0
-                while [ "\$GKE_CLUSTER_STATUS" == "PROVISIONING" ]; do
-                    echo "Cluster \$GKE_CLUSTER is being provisioned, waiting before delete."
-                    sleep 10
+            if gcloud --version > /dev/null 2>&1; then
+                gcloud auth activate-service-account --key-file $CLIENT_SECRET_FILE
+                gcloud config set project $GCP_PROJECT
+                for GKE_CLUSTER in \$(gcloud container clusters list --format='csv[no-heading](name)' --filter="$FILTER"); do
                     GKE_CLUSTER_STATUS=\$(gcloud container clusters list --format='csv[no-heading](status)' --filter="\$GKE_CLUSTER")
-                    let retry+=1
-                    if [ \$retry -ge 60 ]; then
-                        echo "Cluster \$GKE_CLUSTER to delete is being provisioned for too long. Skipping..."
-                        break
-                    fi
+                    retry=0
+                    while [ "\$GKE_CLUSTER_STATUS" == "PROVISIONING" ]; do
+                        echo "Cluster \$GKE_CLUSTER is being provisioned, waiting before delete."
+                        sleep 10
+                        GKE_CLUSTER_STATUS=\$(gcloud container clusters list --format='csv[no-heading](status)' --filter="\$GKE_CLUSTER")
+                        let retry+=1
+                        if [ \$retry -ge 60 ]; then
+                            echo "Cluster \$GKE_CLUSTER to delete is being provisioned for too long. Skipping..."
+                            break
+                        fi
+                    done
+                    gcloud container clusters delete --async --zone $region --quiet \$GKE_CLUSTER || true
                 done
-                gcloud container clusters delete --async --zone $region --quiet \$GKE_CLUSTER || true
-            done
+            fi
         """
    }
 }
